requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
Flask==3.0.0
Flask-MQTT==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.2
paho-mqtt==1.6.1
typing-inspect==0.9.0
typing_extensions==4.8.0
Werkzeug==3.0.1
